iOS/chore : 고민 데이터 샘플 모델 위치 변경
